requirements.txt
# Core dependencies
pyyaml>=6.0.1
pillow>=10.0.0
pydantic>=2.0.0
requests>=2.31.0

# GenAI
openai>=1.12.0

# Environment variables
python-dotenv>=1.0.0

# Phase 2: Localization
deep-translator>=1.11.4

# Phase 2: Brand Compliance
colorthief>=0.2.1

# Optional: Cloud storage (uncomment if needed)
# boto3>=1.34.0
# azure-storage-blob>=12.19.0

# CLI
click>=8.1.7

# Image processing enhancements
numpy>=1.24.0

# Logging
colorama>=0.4.6
